refactor: extract shared diff text renderers to output.diff
