+++ pyproject.toml
@@ -160,6 +160,9 @@ ignore = [
 "src/cja_auto_sdr/diff/snapshot.py" = ["F822", "T201"]  # lazy forwarding + intentional CLI progress output
 "src/cja_auto_sdr/inventory/*.py" = ["F822"] # lazy forwarding modules
 "src/cja_auto_sdr/output/**/*.py" = ["F822"] # lazy forwarding modules
+"src/cja_auto_sdr/output/diff/console.py" = ["RUF001", "RUF003"]  # intentional Unicode box-drawing characters
+"src/cja_auto_sdr/output/diff/grouped.py" = ["RUF001"]  # intentional Unicode arrow (→) in output
+"src/cja_auto_sdr/output/diff/pr_comment.py" = ["RUF001", "RUF003", "PERF401"]  # intentional Unicode emoji (➕➖✏️); PERF401: append-in-slice-loop pattern preserved from generator.py
 "src/cja_auto_sdr/output/sdr/__init__.py" = ["F822", "ERA001", "RUF001", "RUF003"]  # F822: lazy forwarding; ERA001: section header comments; RUF001/3: intentional Unicode symbols
 "src/cja_auto_sdr/pipeline/*.py" = ["F822"]  # lazy forwarding modules
 "tests/**" = ["B", "SIM", "RUF012", "PERF", "DTZ", "S105", "S108", "S110", "S314", "PT017", "PT018", "PT019", "ARG", "T201", "ERA001", "PD011", "FLY002"]
