+++ src/cja_auto_sdr/generator.py
@@ -2605,6 +2605,45 @@ from cja_auto_sdr.api.fetch import EndpointFetchStatus, ParallelAPIFetcher
 # ==================== DATA QUALITY VALIDATION (moved to api/quality.py) ====================
 from cja_auto_sdr.api.quality import DataQualityChecker
 
+# The following functions were moved to cja_auto_sdr.output.sdr:
+# - ExcelFormatCache (class)
+# - apply_excel_formatting
+# - write_excel_output
+# - write_csv_output
+# - write_json_output
+# - write_html_output
+# - write_markdown_output
+# ==================== DIFF COMPARISON OUTPUT WRITERS ====================
+# The following functions were moved to cja_auto_sdr.output.diff:
+# - ANSIColors (class) -> output.diff.common
+# - write_diff_console_output -> output.diff.console
+# - _get_change_symbol, _get_colored_symbol -> output.diff.common
+# - _format_diff_value, _get_change_detail -> output.diff.common
+# - _get_inventory_change_detail -> output.diff.common
+# - _format_side_by_side -> output.diff.console
+# - write_diff_grouped_by_field_output -> output.diff.grouped
+# - write_diff_pr_comment_output -> output.diff.pr_comment
+# - detect_breaking_changes -> output.diff.common
+# - write_diff_markdown_output -> output.diff.markdown
+# - _get_change_emoji -> output.diff.common
+# - _format_markdown_side_by_side -> output.diff.markdown
+from cja_auto_sdr.output.diff import (
+    ANSIColors,
+    _format_diff_value,
+    _format_markdown_side_by_side,
+    _format_side_by_side,
+    _get_change_detail,
+    _get_change_emoji,
+    _get_change_symbol,
+    _get_colored_symbol,
+    _get_inventory_change_detail,
+    detect_breaking_changes,
+    write_diff_console_output,
+    write_diff_grouped_by_field_output,
+    write_diff_markdown_output,
+    write_diff_pr_comment_output,
+)
+
 # ==================== SDR OUTPUT WRITERS (moved to output/sdr/__init__.py) ====================
 from cja_auto_sdr.output.sdr import (
     ExcelFormatCache,
@@ -2615,750 +2654,6 @@ from cja_auto_sdr.output.sdr import (
     write_json_output,
     write_markdown_output,
 )
-
-# The following functions were moved to cja_auto_sdr.output.sdr:
-# - ExcelFormatCache (class)
-# - apply_excel_formatting
-# - write_excel_output
-# - write_csv_output
-# - write_json_output
-# - write_html_output
-# - write_markdown_output
-
-
-# ==================== DIFF COMPARISON OUTPUT WRITERS ====================
-
-
-# ANSIColors is an alias for diff output compatibility
-# It delegates to ConsoleColors but accepts an 'enabled' parameter for explicit control
-class ANSIColors:
-    """ANSI escape codes for colored terminal output.
-
-    This class provides the same functionality as ConsoleColors but with an
-    explicit 'enabled' parameter for cases where color control is needed
-    independent of TTY detection.
-    """
-
-    # Re-export constants from ConsoleColors
-    GREEN = ConsoleColors.GREEN
-    RED = ConsoleColors.RED
-    YELLOW = ConsoleColors.YELLOW
-    CYAN = ConsoleColors.CYAN
-    BOLD = ConsoleColors.BOLD
-    RESET = ConsoleColors.RESET
-    ANSI_ESCAPE = ConsoleColors.ANSI_ESCAPE
-
-    @classmethod
-    def green(cls, text: str, enabled: bool = True) -> str:
-        return f"{cls.GREEN}{text}{cls.RESET}" if enabled else text
-
-    @classmethod
-    def red(cls, text: str, enabled: bool = True) -> str:
-        return f"{cls.RED}{text}{cls.RESET}" if enabled else text
-
-    @classmethod
-    def yellow(cls, text: str, enabled: bool = True) -> str:
-        return f"{cls.YELLOW}{text}{cls.RESET}" if enabled else text
-
-    @classmethod
-    def cyan(cls, text: str, enabled: bool = True) -> str:
-        return f"{cls.CYAN}{text}{cls.RESET}" if enabled else text
-
-    @classmethod
-    def bold(cls, text: str, enabled: bool = True) -> str:
-        return f"{cls.BOLD}{text}{cls.RESET}" if enabled else text
-
-    # Delegate utility methods to ConsoleColors
-    visible_len = ConsoleColors.visible_len
-    rjust = ConsoleColors.rjust
-    ljust = ConsoleColors.ljust
-
-
-def write_diff_console_output(
-    diff_result: DiffResult,
-    changes_only: bool = False,
-    summary_only: bool = False,
-    side_by_side: bool = False,
-    use_color: bool = True,
-) -> str:
-    """
-    Write diff comparison to console with color-coded output.
-
-    Args:
-        diff_result: The DiffResult to output
-        changes_only: Only show changed items (hide unchanged)
-        summary_only: Only show summary statistics
-        side_by_side: Show side-by-side comparison for modified items
-        use_color: Use ANSI color codes in output (default: True)
-
-    Returns:
-        Formatted string for console output
-    """
-    lines = []
-    summary = diff_result.summary
-    meta = diff_result.metadata_diff
-    c = use_color  # Shorthand for color enabled flag
-
-    # Header
-    lines.append("=" * 80)
-    lines.append(ANSIColors.bold("DATA VIEW COMPARISON REPORT", c))
-    lines.append("=" * 80)
-    lines.append(f"{diff_result.source_label}: {meta.source_name} ({meta.source_id})")
-    lines.append(f"{diff_result.target_label}: {meta.target_name} ({meta.target_id})")
-    lines.append(f"Generated: {diff_result.generated_at}")
-    lines.append("=" * 80)
-
-    # Summary table with percentages
-    lines.append("")
-    lines.append(ANSIColors.bold("SUMMARY", c))
-
-    # Build full header labels with data view name and ID
-    src_header = f"{diff_result.source_label}: {meta.source_name} ({meta.source_id})"
-    tgt_header = f"{diff_result.target_label}: {meta.target_name} ({meta.target_id})"
-
-    # Calculate dynamic column widths based on full header lengths
-    src_width = max(8, len(src_header))
-    tgt_width = max(8, len(tgt_header))
-    total_width = 20 + src_width + tgt_width + 10 + 10 + 10 + 12 + 12 + 7  # +7 for spacing
-
-    lines.append(
-        f"{'':20s} {src_header:>{src_width}s} {tgt_header:>{tgt_width}s} {'Added':>10s} {'Removed':>10s} {'Modified':>10s} {'Unchanged':>12s} {'Changed':>12s}",
-    )
-    lines.append("-" * total_width)
-
-    # Metrics row with percentage (using ANSI-aware padding for colored strings)
-    metrics_pct = f"({summary.metrics_change_percent:.1f}%)"
-    added_str = (
-        ANSIColors.green(f"+{summary.metrics_added}", c) if summary.metrics_added else f"+{summary.metrics_added}"
-    )
-    removed_str = (
-        ANSIColors.red(f"-{summary.metrics_removed}", c) if summary.metrics_removed else f"-{summary.metrics_removed}"
-    )
-    modified_str = (
-        ANSIColors.yellow(f"~{summary.metrics_modified}", c)
-        if summary.metrics_modified
-        else f"~{summary.metrics_modified}"
-    )
-    lines.append(
-        f"{'Metrics':20s} {summary.source_metrics_count:{src_width}d} {summary.target_metrics_count:{tgt_width}d} "
-        f"{ANSIColors.rjust(added_str, 10)} {ANSIColors.rjust(removed_str, 10)} {ANSIColors.rjust(modified_str, 10)} {summary.metrics_unchanged:>12d} {metrics_pct:>12s}",
-    )
-
-    # Dimensions row with percentage (using ANSI-aware padding for colored strings)
-    dims_pct = f"({summary.dimensions_change_percent:.1f}%)"
-    added_str = (
-        ANSIColors.green(f"+{summary.dimensions_added}", c)
-        if summary.dimensions_added
-        else f"+{summary.dimensions_added}"
-    )
-    removed_str = (
-        ANSIColors.red(f"-{summary.dimensions_removed}", c)
-        if summary.dimensions_removed
-        else f"-{summary.dimensions_removed}"
-    )
-    modified_str = (
-        ANSIColors.yellow(f"~{summary.dimensions_modified}", c)
-        if summary.dimensions_modified
-        else f"~{summary.dimensions_modified}"
-    )
-    lines.append(
-        f"{'Dimensions':20s} {summary.source_dimensions_count:{src_width}d} {summary.target_dimensions_count:{tgt_width}d} "
-        f"{ANSIColors.rjust(added_str, 10)} {ANSIColors.rjust(removed_str, 10)} {ANSIColors.rjust(modified_str, 10)} {summary.dimensions_unchanged:>12d} {dims_pct:>12s}",
-    )
-
-    # Inventory rows (if present)
-    if summary.has_inventory_changes or (
-        summary.source_calc_metrics_count > 0 or summary.target_calc_metrics_count > 0
-    ):
-        lines.append("-" * total_width)
-        lines.append(ANSIColors.bold("INVENTORY", c))
-
-        # Calculated metrics row
-        if summary.source_calc_metrics_count > 0 or summary.target_calc_metrics_count > 0:
-            added_str = (
-                ANSIColors.green(f"+{summary.calc_metrics_added}", c)
-                if summary.calc_metrics_added
-                else f"+{summary.calc_metrics_added}"
-            )
-            removed_str = (
-                ANSIColors.red(f"-{summary.calc_metrics_removed}", c)
-                if summary.calc_metrics_removed
-                else f"-{summary.calc_metrics_removed}"
-            )
-            modified_str = (
-                ANSIColors.yellow(f"~{summary.calc_metrics_modified}", c)
-                if summary.calc_metrics_modified
-                else f"~{summary.calc_metrics_modified}"
-            )
-            lines.append(
-                f"{'Calc Metrics':20s} {summary.source_calc_metrics_count:{src_width}d} {summary.target_calc_metrics_count:{tgt_width}d} "
-                f"{ANSIColors.rjust(added_str, 10)} {ANSIColors.rjust(removed_str, 10)} {ANSIColors.rjust(modified_str, 10)} {summary.calc_metrics_unchanged:>12d} {'':>12s}",
-            )
-
-        # Segments row
-        if summary.source_segments_count > 0 or summary.target_segments_count > 0:
-            added_str = (
-                ANSIColors.green(f"+{summary.segments_added}", c)
-                if summary.segments_added
-                else f"+{summary.segments_added}"
-            )
-            removed_str = (
-                ANSIColors.red(f"-{summary.segments_removed}", c)
-                if summary.segments_removed
-                else f"-{summary.segments_removed}"
-            )
-            modified_str = (
-                ANSIColors.yellow(f"~{summary.segments_modified}", c)
-                if summary.segments_modified
-                else f"~{summary.segments_modified}"
-            )
-            lines.append(
-                f"{'Segments':20s} {summary.source_segments_count:{src_width}d} {summary.target_segments_count:{tgt_width}d} "
-                f"{ANSIColors.rjust(added_str, 10)} {ANSIColors.rjust(removed_str, 10)} {ANSIColors.rjust(modified_str, 10)} {summary.segments_unchanged:>12d} {'':>12s}",
-            )
-
-    lines.append("-" * total_width)
-
-    if summary_only:
-        lines.append("")
-        if summary.has_changes:
-            lines.append(f"Total changes: {summary.total_changes}")
-            lines.append(f"Summary: {summary.natural_language_summary}")
-        else:
-            lines.append(ANSIColors.green("No differences found.", c))
-        lines.append("=" * 80)
-        return "\n".join(lines)
-
-    # Get changes for both metrics and dimensions
-    metric_changes = [d for d in diff_result.metric_diffs if d.change_type != ChangeType.UNCHANGED]
-    dim_changes = [d for d in diff_result.dimension_diffs if d.change_type != ChangeType.UNCHANGED]
-
-    # Calculate global max ID width for consistent alignment across both sections
-    all_changes = metric_changes + dim_changes
-    global_max_id_len = max((len(d.id) for d in all_changes), default=0)
-
-    # Metrics changes
-    if metric_changes or not changes_only:
-        lines.append("")
-        change_count = len(metric_changes)
-        lines.append(ANSIColors.bold(f"METRICS CHANGES ({change_count})", c))
-        if metric_changes:
-            for diff in metric_changes:
-                colored_symbol = _get_colored_symbol(diff.change_type, c)
-                lines.append(f'  [{colored_symbol}] {diff.id:{global_max_id_len}s} "{diff.name}"')
-                if side_by_side and diff.change_type == ChangeType.MODIFIED:
-                    # Side-by-side view for modified items
-                    sbs_lines = _format_side_by_side(diff, diff_result.source_label, diff_result.target_label)
-                    lines.extend(sbs_lines)
-                else:
-                    detail = _get_change_detail(diff)
-                    if detail:
-                        lines.append(f"      {detail}")
-        else:
-            lines.append("  No changes")
-
-    # Dimensions changes
-    if dim_changes or not changes_only:
-        lines.append("")
-        change_count = len(dim_changes)
-        lines.append(ANSIColors.bold(f"DIMENSIONS CHANGES ({change_count})", c))
-        if dim_changes:
-            for diff in dim_changes:
-                colored_symbol = _get_colored_symbol(diff.change_type, c)
-                lines.append(f'  [{colored_symbol}] {diff.id:{global_max_id_len}s} "{diff.name}"')
-                if side_by_side and diff.change_type == ChangeType.MODIFIED:
-                    # Side-by-side view for modified items
-                    sbs_lines = _format_side_by_side(diff, diff_result.source_label, diff_result.target_label)
-                    lines.extend(sbs_lines)
-                else:
-                    detail = _get_change_detail(diff)
-                    if detail:
-                        lines.append(f"      {detail}")
-        else:
-            lines.append("  No changes")
-
-    # Inventory changes (if present)
-    if diff_result.has_inventory_diffs:
-        lines.append("")
-        lines.append("-" * 80)
-        lines.append(ANSIColors.bold("INVENTORY CHANGES", c))
-        lines.append("-" * 80)
-
-        # Calculated metrics inventory changes
-        if diff_result.calc_metrics_diffs is not None:
-            calc_changes = [d for d in diff_result.calc_metrics_diffs if d.change_type != ChangeType.UNCHANGED]
-            lines.append("")
-            lines.append(ANSIColors.bold(f"CALCULATED METRICS ({len(calc_changes)} changes)", c))
-            if calc_changes:
-                for diff in calc_changes:
-                    colored_symbol = _get_colored_symbol(diff.change_type, c)
-                    lines.append(f'  [{colored_symbol}] {diff.id} "{diff.name}"')
-                    if diff.change_type == ChangeType.MODIFIED and diff.changed_fields:
-                        detail = _get_inventory_change_detail(diff)
-                        if detail:
-                            lines.append(f"      {detail}")
-            else:
-                lines.append("  No changes")
-
-        # Segments inventory changes
-        if diff_result.segments_diffs is not None:
-            seg_changes = [d for d in diff_result.segments_diffs if d.change_type != ChangeType.UNCHANGED]
-            lines.append("")
-            lines.append(ANSIColors.bold(f"SEGMENTS ({len(seg_changes)} changes)", c))
-            if seg_changes:
-                for diff in seg_changes:
-                    colored_symbol = _get_colored_symbol(diff.change_type, c)
-                    lines.append(f'  [{colored_symbol}] {diff.id} "{diff.name}"')
-                    if diff.change_type == ChangeType.MODIFIED and diff.changed_fields:
-                        detail = _get_inventory_change_detail(diff)
-                        if detail:
-                            lines.append(f"      {detail}")
-            else:
-                lines.append("  No changes")
-
-    # Footer with total summary
-    lines.append("")
-    lines.append("=" * 80)
-    if summary.has_changes:
-        # Build color-coded total summary line
-        total_parts = []
-        if summary.total_added:
-            total_parts.append(ANSIColors.green(f"{summary.total_added} added", c))
-        if summary.total_removed:
-            total_parts.append(ANSIColors.red(f"{summary.total_removed} removed", c))
-        if summary.total_modified:
-            total_parts.append(ANSIColors.yellow(f"{summary.total_modified} modified", c))
-        total_line = ", ".join(total_parts)
-        lines.append(ANSIColors.bold(f"Total: {total_line}", c))
-        lines.append(
-            f"  Metrics: {summary.metrics_added} added, {summary.metrics_removed} removed, {summary.metrics_modified} modified",
-        )
-        lines.append(
-            f"  Dimensions: {summary.dimensions_added} added, {summary.dimensions_removed} removed, {summary.dimensions_modified} modified",
-        )
-        # Add inventory summary lines if present
-        if summary.has_inventory_changes:
-            if summary.source_calc_metrics_count > 0 or summary.target_calc_metrics_count > 0:
-                lines.append(
-                    f"  Calc Metrics: {summary.calc_metrics_added} added, {summary.calc_metrics_removed} removed, {summary.calc_metrics_modified} modified",
-                )
-            if summary.source_segments_count > 0 or summary.target_segments_count > 0:
-                lines.append(
-                    f"  Segments: {summary.segments_added} added, {summary.segments_removed} removed, {summary.segments_modified} modified",
-                )
-    else:
-        lines.append(ANSIColors.green("✓ No differences found", c))
-    lines.append("=" * 80)
-
-    return "\n".join(lines)
-
-
-def _get_change_symbol(change_type: ChangeType) -> str:
-    """Get symbol for change type"""
-    symbols = {ChangeType.ADDED: "+", ChangeType.REMOVED: "-", ChangeType.MODIFIED: "~", ChangeType.UNCHANGED: " "}
-    return symbols.get(change_type, "?")
-
-
-def _get_colored_symbol(change_type: ChangeType, use_color: bool = True) -> str:
-    """Get color-coded symbol for change type"""
-    symbol = _get_change_symbol(change_type)
-    if not use_color:
-        return symbol
-    if change_type == ChangeType.ADDED:
-        return ANSIColors.green(symbol, use_color)
-    if change_type == ChangeType.REMOVED:
-        return ANSIColors.red(symbol, use_color)
-    if change_type == ChangeType.MODIFIED:
-        return ANSIColors.yellow(symbol, use_color)
-    return symbol
-
-
-def _format_diff_value(val: Any, truncate: bool = True, max_len: int = 30) -> str:
-    """Format a value for diff display, handling None and NaN."""
-    if val is None:
-        return "(empty)"
-    try:
-        if pd.isna(val):
-            return "(empty)"
-    except (TypeError, ValueError):
-        pass
-    result = str(val)
-    if truncate and len(result) > max_len:
-        result = result[:max_len]
-    return result
-
-
-def _get_change_detail(diff: ComponentDiff, truncate: bool = True) -> str:
-    """Get detail string for a component diff"""
-    if diff.change_type == ChangeType.MODIFIED and diff.changed_fields:
-        changes = []
-        for field, (old_val, new_val) in diff.changed_fields.items():
-            old_str = _format_diff_value(old_val, truncate)
-            new_str = _format_diff_value(new_val, truncate)
-            changes.append(f"{field}: '{old_str}' -> '{new_str}'")
-        return "; ".join(changes)
-    return ""
-
-
-def _get_inventory_change_detail(diff: InventoryItemDiff, truncate: bool = True) -> str:
-    """Get detail string for an inventory item diff"""
-    if diff.change_type == ChangeType.MODIFIED and diff.changed_fields:
-        changes = []
-        for field, (old_val, new_val) in diff.changed_fields.items():
-            old_str = _format_diff_value(old_val, truncate)
-            new_str = _format_diff_value(new_val, truncate)
-            changes.append(f"{field}: '{old_str}' -> '{new_str}'")
-        return "; ".join(changes)
-    return ""
-
-
-def _format_side_by_side(
-    diff: ComponentDiff,
-    source_label: str,
-    target_label: str,
-    col_width: int = 35,
-    max_col_width: int = 60,
-) -> list[str]:
-    """
-    Format a component diff as a side-by-side comparison table.
-
-    Args:
-        diff: The ComponentDiff to format
-        source_label: Label for source side
-        target_label: Label for target side
-        col_width: Base width of each column
-        max_col_width: Maximum width of each column (text will wrap)
-
-    Returns:
-        List of formatted lines for the side-by-side view
-    """
-    lines = []
-    if diff.change_type != ChangeType.MODIFIED or not diff.changed_fields:
-        return lines
-
-    # Pre-compute all display strings
-    field_displays = []
-    for field_name, (old_val, new_val) in diff.changed_fields.items():
-        old_str = _format_diff_value(old_val, truncate=False)
-        new_str = _format_diff_value(new_val, truncate=False)
-        old_display = f"{field_name}: {old_str}"
-        new_display = f"{field_name}: {new_str}"
-        field_displays.append((old_display, new_display))
-
-    # Calculate column width: expand to fit content but cap at max_col_width
-    col_width = max(col_width, len(source_label) + 2, len(target_label) + 2)
-    for old_display, new_display in field_displays:
-        col_width = max(col_width, min(len(old_display), max_col_width), min(len(new_display), max_col_width))
-    col_width = min(col_width, max_col_width)
-
-    # Header for this component
-    lines.append(f"    ┌{'─' * (col_width + 2)}┬{'─' * (col_width + 2)}┐")
-    lines.append(f"    │ {source_label:<{col_width}} │ {target_label:<{col_width}} │")
-    lines.append(f"    ├{'─' * (col_width + 2)}┼{'─' * (col_width + 2)}┤")
-
-    # Changed fields with text wrapping
-    for old_display, new_display in field_displays:
-        # Wrap each side independently
-        old_wrapped = textwrap.wrap(old_display, width=col_width) or [""]
-        new_wrapped = textwrap.wrap(new_display, width=col_width) or [""]
-
-        # Pad to same number of lines
-        max_lines = max(len(old_wrapped), len(new_wrapped))
-        old_wrapped.extend([""] * (max_lines - len(old_wrapped)))
-        new_wrapped.extend([""] * (max_lines - len(new_wrapped)))
-
-        # Output each line
-        for old_line, new_line in zip(old_wrapped, new_wrapped, strict=True):
-            lines.append(f"    │ {old_line:<{col_width}} │ {new_line:<{col_width}} │")
-
-    lines.append(f"    └{'─' * (col_width + 2)}┴{'─' * (col_width + 2)}┘")
-
-    return lines
-
-
-def write_diff_grouped_by_field_output(diff_result: DiffResult, use_color: bool = True, limit: int = 10) -> str:
-    """
-    Write diff output grouped by changed field instead of by component.
-
-    Args:
-        diff_result: The DiffResult to output
-        use_color: Use ANSI color codes
-        limit: Max items per section (0 = unlimited)
-
-    Returns:
-        Formatted string for console output
-    """
-    lines = []
-    summary = diff_result.summary
-    meta = diff_result.metadata_diff
-    c = use_color
-
-    # Header
-    lines.append("=" * 80)
-    lines.append(ANSIColors.bold("DATA VIEW COMPARISON - GROUPED BY FIELD", c))
-    lines.append("=" * 80)
-    lines.append(f"{diff_result.source_label}: {meta.source_name}")
-    lines.append(f"{diff_result.target_label}: {meta.target_name}")
-    lines.append(f"Generated: {diff_result.generated_at}")
-    lines.append("=" * 80)
-
-    # Collect all changed fields across all components
-    field_changes: dict[str, list[tuple[str, str, Any, Any]]] = {}  # field -> [(id, name, old, new), ...]
-
-    # Also track breaking changes (type or schemaPath changes)
-    breaking_changes = []
-
-    all_diffs = diff_result.metric_diffs + diff_result.dimension_diffs
-    for diff in all_diffs:
-        if diff.change_type == ChangeType.MODIFIED and diff.changed_fields:
-            for field, (old_val, new_val) in diff.changed_fields.items():
-                if field not in field_changes:
-                    field_changes[field] = []
-                field_changes[field].append((diff.id, diff.name, old_val, new_val))
-
-                # Track breaking changes
-                if field in ("type", "schemaPath"):
-                    breaking_changes.append((diff.id, diff.name, field, old_val, new_val))
-
-    # Summary
-    lines.append("")
-    lines.append(ANSIColors.bold("SUMMARY", c))
-    lines.append(f"Total components changed: {summary.total_changes}")
-    lines.append(f"  Added: {ANSIColors.green(str(summary.metrics_added + summary.dimensions_added), c)}")
-    lines.append(f"  Removed: {ANSIColors.red(str(summary.metrics_removed + summary.dimensions_removed), c)}")
-    lines.append(f"  Modified: {ANSIColors.yellow(str(summary.metrics_modified + summary.dimensions_modified), c)}")
-    lines.append(f"Fields with changes: {len(field_changes)}")
-
-    # Breaking changes warning
-    if breaking_changes:
-        lines.append("")
-        lines.append(ANSIColors.red("⚠  BREAKING CHANGES DETECTED", c))
-        lines.append("-" * 40)
-        for comp_id, _comp_name, field, old_val, new_val in breaking_changes:
-            lines.append(f"  {comp_id}: {field} changed")
-            lines.append(
-                f"    '{_format_diff_value(old_val, truncate=False)}' → '{_format_diff_value(new_val, truncate=False)}'",
-            )
-
-    # Group by field
-    lines.append("")
-    lines.append(ANSIColors.bold("CHANGES BY FIELD", c))
-    lines.append("-" * 80)
-
-    for field in sorted(field_changes.keys()):
-        changes = field_changes[field]
-        lines.append("")
-        lines.append(f"{ANSIColors.cyan(field, c)} ({len(changes)} component{'s' if len(changes) != 1 else ''}):")
-
-        items_to_show = changes if limit == 0 else changes[:limit]
-        for comp_id, _comp_name, old_val, new_val in items_to_show:
-            old_str = _format_diff_value(old_val, truncate=True)
-            new_str = _format_diff_value(new_val, truncate=True)
-            lines.append(f"  {comp_id}: '{old_str}' → '{new_str}'")
-
-        if limit > 0 and len(changes) > limit:
-            lines.append(f"  ... and {len(changes) - limit} more")
-
-    # Added/removed summary
-    added = [d for d in all_diffs if d.change_type == ChangeType.ADDED]
-    removed = [d for d in all_diffs if d.change_type == ChangeType.REMOVED]
-
-    if added:
-        lines.append("")
-        lines.append(ANSIColors.green(f"ADDED ({len(added)})", c))
-        added_to_show = added if limit == 0 else added[:limit]
-        lines.extend(f"  [+] {diff.id}" for diff in added_to_show)
-        if limit > 0 and len(added) > limit:
-            lines.append(f"  ... and {len(added) - limit} more")
-
-    if removed:
-        lines.append("")
-        lines.append(ANSIColors.red(f"REMOVED ({len(removed)})", c))
-        removed_to_show = removed if limit == 0 else removed[:limit]
-        lines.extend(f"  [-] {diff.id}" for diff in removed_to_show)
-        if limit > 0 and len(removed) > limit:
-            lines.append(f"  ... and {len(removed) - limit} more")
-
-    lines.append("")
-    lines.append("=" * 80)
-    lines.append(ANSIColors.cyan(f"Summary: {summary.natural_language_summary}", c))
-    lines.append("=" * 80)
-
-    return "\n".join(lines)
-
-
-def write_diff_pr_comment_output(diff_result: DiffResult, _changes_only: bool = False) -> str:
-    """
-    Write diff output in GitHub/GitLab PR comment format with collapsible details.
-
-    Args:
-        diff_result: The DiffResult to output
-        changes_only: Only include changed items
-
-    Returns:
-        Markdown formatted string optimized for PR comments
-    """
-    lines = []
-    summary = diff_result.summary
-
-    # Header
-    lines.append("### 📊 Data View Comparison")
-    lines.append("")
-    lines.append(f"**{diff_result.source_label}** → **{diff_result.target_label}**")
-    lines.append("")
-
-    # Summary table
-    lines.append("| Component | Source | Target | Added | Removed | Modified | Unchanged | Changed |")
-    lines.append("|-----------|-------:|-------:|------:|--------:|---------:|----------:|--------:|")
-    lines.append(
-        f"| Metrics | {summary.source_metrics_count} | {summary.target_metrics_count} | "
-        f"+{summary.metrics_added} | -{summary.metrics_removed} | ~{summary.metrics_modified} | "
-        f"{summary.metrics_unchanged} | {summary.metrics_change_percent:.1f}% |",
-    )
-    lines.append(
-        f"| Dimensions | {summary.source_dimensions_count} | {summary.target_dimensions_count} | "
-        f"+{summary.dimensions_added} | -{summary.dimensions_removed} | ~{summary.dimensions_modified} | "
-        f"{summary.dimensions_unchanged} | {summary.dimensions_change_percent:.1f}% |",
-    )
-    lines.append("")
-
-    # Breaking changes warning
-    breaking_changes = []
-    all_diffs = diff_result.metric_diffs + diff_result.dimension_diffs
-    for diff in all_diffs:
-        if diff.change_type == ChangeType.MODIFIED and diff.changed_fields:
-            for field in diff.changed_fields:
-                if field in ("type", "schemaPath"):
-                    old_val, new_val = diff.changed_fields[field]
-                    breaking_changes.append((diff.id, field, old_val, new_val))
-
-    if breaking_changes:
-        lines.append("#### ⚠ Breaking Changes Detected")
-        lines.append("")
-        lines.append("| Component | Field | Before | After |")
-        lines.append("|-----------|-------|--------|-------|")
-        for comp_id, field, old_val, new_val in breaking_changes[:10]:
-            lines.append(
-                f"| `{comp_id}` | {field} | `{_format_diff_value(old_val, truncate=False)}` | `{_format_diff_value(new_val, truncate=False)}` |",
-            )
-        if len(breaking_changes) > 10:
-            lines.append(f"| ... | | | +{len(breaking_changes) - 10} more |")
-        lines.append("")
-
-    # Natural language summary
-    lines.append(f"**Summary:** {summary.natural_language_summary}")
-    lines.append("")
-
-    # Collapsible details
-    metric_changes = [d for d in diff_result.metric_diffs if d.change_type != ChangeType.UNCHANGED]
-    dim_changes = [d for d in diff_result.dimension_diffs if d.change_type != ChangeType.UNCHANGED]
-
-    if metric_changes:
-        lines.append("<details>")
-        lines.append(f"<summary>📈 Metrics Changes ({len(metric_changes)})</summary>")
-        lines.append("")
-        lines.append("| Change | ID | Name |")
-        lines.append("|--------|----|----- |")
-        for diff in metric_changes[:25]:
-            symbol = {ChangeType.ADDED: "➕", ChangeType.REMOVED: "➖", ChangeType.MODIFIED: "✏️"}.get(
-                diff.change_type,
-                "",
-            )
-            lines.append(f"| {symbol} | `{diff.id}` | {diff.name} |")
-        if len(metric_changes) > 25:
-            lines.append(f"| ... | | +{len(metric_changes) - 25} more |")
-        lines.append("")
-        lines.append("</details>")
-        lines.append("")
-
-    if dim_changes:
-        lines.append("<details>")
-        lines.append(f"<summary>📏 Dimensions Changes ({len(dim_changes)})</summary>")
-        lines.append("")
-        lines.append("| Change | ID | Name |")
-        lines.append("|--------|----|----- |")
-        for diff in dim_changes[:25]:
-            symbol = {ChangeType.ADDED: "➕", ChangeType.REMOVED: "➖", ChangeType.MODIFIED: "✏️"}.get(
-                diff.change_type,
-                "",
-            )
-            lines.append(f"| {symbol} | `{diff.id}` | {diff.name} |")
-        if len(dim_changes) > 25:
-            lines.append(f"| ... | | +{len(dim_changes) - 25} more |")
-        lines.append("")
-        lines.append("</details>")
-        lines.append("")
-
-    # Footer
-    lines.append("---")
-    lines.append(f"*Generated by CJA SDR Generator v{diff_result.tool_version}*")
-
-    return "\n".join(lines)
-
-
-def detect_breaking_changes(diff_result: DiffResult) -> list[dict[str, Any]]:
-    """
-    Detect breaking changes in a diff result.
-
-    Breaking changes include:
-    - Changes to 'type' field (data type changes)
-    - Changes to 'schemaPath' field (schema mapping changes)
-    - Removal of existing components
-
-    Args:
-        diff_result: The DiffResult to analyze
-
-    Returns:
-        List of breaking change dictionaries with details
-    """
-    breaking_changes = []
-
-    all_diffs = diff_result.metric_diffs + diff_result.dimension_diffs
-
-    for diff in all_diffs:
-        # Removed components are breaking
-        if diff.change_type == ChangeType.REMOVED:
-            breaking_changes.append(
-                {
-                    "component_id": diff.id,
-                    "component_name": diff.name,
-                    "change_type": "removed",
-                    "severity": "high",
-                    "description": f"Component '{diff.name}' was removed",
-                },
-            )
-
-        # Check for type or schema changes
-        elif diff.change_type == ChangeType.MODIFIED and diff.changed_fields:
-            for field, (old_val, new_val) in diff.changed_fields.items():
-                if field == "type":
-                    breaking_changes.append(
-                        {
-                            "component_id": diff.id,
-                            "component_name": diff.name,
-                            "change_type": "type_changed",
-                            "field": field,
-                            "old_value": old_val,
-                            "new_value": new_val,
-                            "severity": "high",
-                            "description": f"Data type changed from '{_format_diff_value(old_val, truncate=False)}' to '{_format_diff_value(new_val, truncate=False)}'",
-                        },
-                    )
-                elif field == "schemaPath":
-                    breaking_changes.append(
-                        {
-                            "component_id": diff.id,
-                            "component_name": diff.name,
-                            "change_type": "schema_changed",
-                            "field": field,
-                            "old_value": old_val,
-                            "new_value": new_val,
-                            "severity": "medium",
-                            "description": f"Schema path changed from '{_format_diff_value(old_val, truncate=False)}' to '{_format_diff_value(new_val, truncate=False)}'",
-                        },
-                    )
-
-    return breaking_changes
 
 
 def write_diff_json_output(
@@ -3505,231 +2800,6 @@ def write_diff_json_output(
     except (OSError, TypeError, ValueError) as e:
         logger.error(_format_error_msg("creating diff JSON file", error=e))
         raise
-
-
-def write_diff_markdown_output(
-    diff_result: DiffResult,
-    base_filename: str,
-    output_dir: str | Path,
-    logger: logging.Logger,
-    changes_only: bool = False,
-    side_by_side: bool = False,
-) -> str:
-    """
-    Write diff comparison to Markdown format.
-
-    Args:
-        diff_result: The DiffResult to output
-        base_filename: Base filename without extension
-        output_dir: Output directory path
-        logger: Logger instance
-        changes_only: Only include changed items
-        side_by_side: Show side-by-side comparison for modified items
-
-    Returns:
-        Path to Markdown output file
-    """
-    try:
-        logger.info("Generating diff Markdown output...")
-
-        summary = diff_result.summary
-        meta = diff_result.metadata_diff
-        md_parts = []
-
-        # Title
-        md_parts.append("# Data View Comparison Report\n")
-
-        # Metadata
-        md_parts.append("## Comparison Details\n")
-        md_parts.append(f"**{diff_result.source_label}:** {meta.source_name} (`{meta.source_id}`)")
-        md_parts.append(f"**{diff_result.target_label}:** {meta.target_name} (`{meta.target_id}`)")
-        md_parts.append(f"**Generated:** {diff_result.generated_at}")
-        md_parts.append(f"**Tool Version:** {diff_result.tool_version}\n")
-
-        # Summary table
-        md_parts.append("## Summary\n")
-        md_parts.append(
-            f"| Component | {diff_result.source_label} | {diff_result.target_label} | Added | Removed | Modified | Unchanged | Changed |",
-        )
-        md_parts.append("| --- | ---: | ---: | ---: | ---: | ---: | ---: | ---: |")
-        md_parts.append(
-            f"| Metrics | {summary.source_metrics_count} | {summary.target_metrics_count} | "
-            f"+{summary.metrics_added} | -{summary.metrics_removed} | ~{summary.metrics_modified} | "
-            f"{summary.metrics_unchanged} | {summary.metrics_change_percent:.1f}% |",
-        )
-        md_parts.append(
-            f"| Dimensions | {summary.source_dimensions_count} | {summary.target_dimensions_count} | "
-            f"+{summary.dimensions_added} | -{summary.dimensions_removed} | ~{summary.dimensions_modified} | "
-            f"{summary.dimensions_unchanged} | {summary.dimensions_change_percent:.1f}% |",
-        )
-
-        # Add inventory rows to summary if present
-        if diff_result.has_inventory_diffs:
-            if summary.source_calc_metrics_count > 0 or summary.target_calc_metrics_count > 0:
-                md_parts.append(
-                    f"| **Calc Metrics** | {summary.source_calc_metrics_count} | {summary.target_calc_metrics_count} | "
-                    f"+{summary.calc_metrics_added} | -{summary.calc_metrics_removed} | ~{summary.calc_metrics_modified} | "
-                    f"{summary.calc_metrics_unchanged} | - |",
-                )
-            if summary.source_segments_count > 0 or summary.target_segments_count > 0:
-                md_parts.append(
-                    f"| **Segments** | {summary.source_segments_count} | {summary.target_segments_count} | "
-                    f"+{summary.segments_added} | -{summary.segments_removed} | ~{summary.segments_modified} | "
-                    f"{summary.segments_unchanged} | - |",
-                )
-        md_parts.append("")
-
-        if not summary.has_changes:
-            md_parts.append("**✓ No differences found.**\n")
-        else:
-            md_parts.append(f"**Total: {summary.total_summary}**\n")
-
-        # Metrics changes
-        metric_changes = [d for d in diff_result.metric_diffs if d.change_type != ChangeType.UNCHANGED]
-        if metric_changes or not changes_only:
-            md_parts.append("## Metrics Changes\n")
-            if metric_changes:
-                md_parts.append("| Status | ID | Name | Details |")
-                md_parts.append("| --- | --- | --- | --- |")
-                for diff in metric_changes:
-                    symbol = _get_change_emoji(diff.change_type)
-                    detail = _get_change_detail(diff).replace("|", "\\|")
-                    md_parts.append(f"| {symbol} | `{diff.id}` | {diff.name} | {detail} |")
-
-                # Add side-by-side detail for modified items
-                if side_by_side:
-                    modified = [d for d in metric_changes if d.change_type == ChangeType.MODIFIED]
-                    if modified:
-                        md_parts.append("\n### Modified Metrics - Side by Side\n")
-                        for diff in modified:
-                            md_parts.extend(
-                                _format_markdown_side_by_side(diff, diff_result.source_label, diff_result.target_label),
-                            )
-            else:
-                md_parts.append("*No changes*")
-            md_parts.append("")
-
-        # Dimensions changes
-        dim_changes = [d for d in diff_result.dimension_diffs if d.change_type != ChangeType.UNCHANGED]
-        if dim_changes or not changes_only:
-            md_parts.append("## Dimensions Changes\n")
-            if dim_changes:
-                md_parts.append("| Status | ID | Name | Details |")
-                md_parts.append("| --- | --- | --- | --- |")
-                for diff in dim_changes:
-                    symbol = _get_change_emoji(diff.change_type)
-                    detail = _get_change_detail(diff).replace("|", "\\|")
-                    md_parts.append(f"| {symbol} | `{diff.id}` | {diff.name} | {detail} |")
-
-                # Add side-by-side detail for modified items
-                if side_by_side:
-                    modified = [d for d in dim_changes if d.change_type == ChangeType.MODIFIED]
-                    if modified:
-                        md_parts.append("\n### Modified Dimensions - Side by Side\n")
-                        for diff in modified:
-                            md_parts.extend(
-                                _format_markdown_side_by_side(diff, diff_result.source_label, diff_result.target_label),
-                            )
-            else:
-                md_parts.append("*No changes*")
-            md_parts.append("")
-
-        # Inventory changes (if present)
-        if diff_result.has_inventory_diffs:
-            md_parts.append("---\n")
-            md_parts.append("# Inventory Changes\n")
-
-            # Calculated metrics inventory changes
-            if diff_result.calc_metrics_diffs is not None:
-                calc_changes = [d for d in diff_result.calc_metrics_diffs if d.change_type != ChangeType.UNCHANGED]
-                md_parts.append(f"## Calculated Metrics Changes ({len(calc_changes)})\n")
-                if calc_changes:
-                    md_parts.append("| Status | ID | Name | Details |")
-                    md_parts.append("| --- | --- | --- | --- |")
-                    for diff in calc_changes:
-                        symbol = _get_change_emoji(diff.change_type)
-                        detail = _get_inventory_change_detail(diff).replace("|", "\\|")
-                        md_parts.append(f"| {symbol} | `{diff.id}` | {diff.name} | {detail} |")
-                else:
-                    md_parts.append("*No changes*")
-                md_parts.append("")
-
-            # Segments inventory changes
-            if diff_result.segments_diffs is not None:
-                seg_changes = [d for d in diff_result.segments_diffs if d.change_type != ChangeType.UNCHANGED]
-                md_parts.append(f"## Segments Changes ({len(seg_changes)})\n")
-                if seg_changes:
-                    md_parts.append("| Status | ID | Name | Details |")
-                    md_parts.append("| --- | --- | --- | --- |")
-                    for diff in seg_changes:
-                        symbol = _get_change_emoji(diff.change_type)
-                        detail = _get_inventory_change_detail(diff).replace("|", "\\|")
-                        md_parts.append(f"| {symbol} | `{diff.id}` | {diff.name} | {detail} |")
-                else:
-                    md_parts.append("*No changes*")
-                md_parts.append("")
-
-        md_parts.append("---")
-        md_parts.append("*Generated by CJA Auto SDR Generator*")
-
-        markdown_file = os.path.join(output_dir, f"{base_filename}.md")
-        with open(markdown_file, "w", encoding="utf-8") as f:
-            f.write("\n".join(md_parts))
-
-        logger.info(f"Diff Markdown file created: {markdown_file}")
-        return markdown_file
-
-    except (OSError, KeyError, TypeError, ValueError) as e:
-        logger.error(_format_error_msg("creating diff Markdown file", error=e))
-        raise
-
-
-def _get_change_emoji(change_type: ChangeType) -> str:
-    """Get emoji for change type"""
-    emojis = {ChangeType.ADDED: "+", ChangeType.REMOVED: "-", ChangeType.MODIFIED: "~", ChangeType.UNCHANGED: ""}
-    return emojis.get(change_type, "")
-
-
-def _format_markdown_side_by_side(diff: ComponentDiff, source_label: str, target_label: str) -> list[str]:
-    """
-    Format a component diff as a side-by-side markdown table.
-
-    Args:
-        diff: The ComponentDiff to format
-        source_label: Label for source side
-        target_label: Label for target side
-
-    Returns:
-        List of markdown lines for the side-by-side view
-    """
-    lines = []
-    if diff.change_type != ChangeType.MODIFIED or not diff.changed_fields:
-        return lines
-
-    # Component header
-    lines.append(f"\n**`{diff.id}`** - {diff.name}\n")
-
-    # Side-by-side table
-    lines.append(f"| Field | {source_label} | {target_label} |")
-    lines.append("| --- | --- | --- |")
-
-    for field_name, (old_val, new_val) in diff.changed_fields.items():
-        old_formatted = _format_diff_value(old_val, truncate=False)
-        new_formatted = _format_diff_value(new_val, truncate=False)
-        # Use italic for empty values in markdown
-        old_str = "*(empty)*" if old_formatted == "(empty)" else old_formatted.replace("|", "\\|")
-        new_str = "*(empty)*" if new_formatted == "(empty)" else new_formatted.replace("|", "\\|")
-
-        # Truncate very long values
-        if len(old_str) > 50:
-            old_str = old_str[:47] + "..."
-        if len(new_str) > 50:
-            new_str = new_str[:47] + "..."
-
-        lines.append(f"| `{field_name}` | {old_str} | {new_str} |")
-
-    lines.append("")
-    return lines
 
 
 def write_diff_html_output(
